correcion de errores 3
